requirements.txt
beautifulsoup4==4.12.3
pick==2.3.2
Pillow==10.3.0
Requests==2.32.3
rich==13.7.1
